+++ src/AudioOutputsService.ts
@@ -35,15 +35,15 @@ type AudioOutputsServiceSource = {
 class AudioOutputsService extends EventEmitter<AudioOutputsServiceEvents> {
   private audioContext: AudioContext | undefined;
 
-  private gainNode: GainNode | undefined;
+  private muteGainNode: GainNode | null = null;
+  private volumeGainNode: GainNode | null = null;
+  private analyserNode: AnalyserNode | null = null;
 
-  private analyserNode: AnalyserNode | undefined;
+  public normalVolume = 1;
 
   private currentSources: AudioOutputsServiceSource[] = [];
 
   private debugLogFunction: (message: string) => void;
-
-  public currentVolume = 1;
 
   constructor(debugLogFunction: (message: string) => void) {
     super();
@@ -65,14 +65,20 @@ class AudioOutputsService extends EventEmitter<AudioOutputsServiceEvents> {
     this.audioContext = new AudioContextClass();
 
     // Create and store the gain node.
-    this.gainNode = this.audioContext.createGain();
-    this.gainNode.gain.value = this.currentVolume;
+    this.muteGainNode = this.audioContext.createGain();
+    this.volumeGainNode = this.audioContext.createGain();
 
     // Create and store the analyser node
     this.analyserNode = this.audioContext.createAnalyser();
 
-    // Connect the gain node to the analyser node and then to the destination
-    this.gainNode.connect(this.analyserNode);
+    this.muteGainNode.gain.setValueAtTime(1, this.audioContext.currentTime);
+    this.volumeGainNode.gain.setValueAtTime(
+      this.normalVolume,
+      this.audioContext.currentTime,
+    );
+
+    this.volumeGainNode.connect(this.muteGainNode);
+    this.muteGainNode.connect(this.analyserNode);
     this.analyserNode.connect(this.audioContext.destination);
 
     return this.audioContext;
@@ -115,13 +121,12 @@ class AudioOutputsService extends EventEmitter<AudioOutputsServiceEvents> {
 
     const audioContext = this.getAudioContext();
 
-    // Assert that gainNode is not undefined
-    if (!this.gainNode) {
-      throw new Error("GainNode is not initialized.");
+    if (!this.volumeGainNode) {
+      throw new Error("volumeGainNode is not initialized.");
     }
 
     const source = audioContext.createBufferSource();
-    source.connect(this.gainNode);
+    source.connect(this.volumeGainNode);
     source.buffer = await new Promise((resolve, reject): void => {
       audioContext.decodeAudioData(audio, resolve, reject);
     });
@@ -154,23 +159,55 @@ class AudioOutputsService extends EventEmitter<AudioOutputsServiceEvents> {
     });
   };
 
-  public setVolume = (volume: number): void => {
-    this.debugLogFunction(`AudioOutputsService setVolume ${volume}`);
-    if (!this.gainNode) return;
+  public setNormalVolume = (volume: number): void => {
+    this.debugLogFunction(`AudioOutputsService setNormalVolume ${volume}`);
+    if (!this.volumeGainNode || !this.audioContext) return;
 
     // Clamp the volume to the range [0, 1]
     const clampedVolume = Math.max(0, Math.min(1, volume));
 
-    // Store the volume value
-    this.currentVolume = clampedVolume;
+    // record volume on a variable in case volume is requested before ramp has finished
+    this.normalVolume = clampedVolume;
 
-    // Set the gain value
-    this.gainNode.gain.value = clampedVolume;
+    // smooth ramp to new value
+    this.volumeGainNode.gain.setValueAtTime(
+      this.volumeGainNode.gain.value,
+      this.audioContext.currentTime,
+    );
+    this.volumeGainNode.gain.linearRampToValueAtTime(
+      clampedVolume,
+      this.audioContext.currentTime + 0.1,
+    );
   };
 
-  public getVolume = (): number => {
-    this.debugLogFunction("AudioOutputsService getVolume");
-    return this.currentVolume;
+  public beginMuting = (): void => {
+    this.debugLogFunction(`AudioOutputsService beginMuting`);
+    if (!this.muteGainNode || !this.audioContext) return;
+
+    // Fade out quickly
+    this.muteGainNode.gain.setValueAtTime(
+      this.muteGainNode.gain.value,
+      this.audioContext.currentTime,
+    );
+    this.muteGainNode.gain.linearRampToValueAtTime(
+      0,
+      this.audioContext.currentTime + 0.05,
+    );
+  };
+
+  public endMuting = (): void => {
+    this.debugLogFunction(`AudioOutputsService endMuting`);
+    if (!this.muteGainNode || !this.audioContext) return;
+
+    // Fade in very quickly
+    this.muteGainNode.gain.setValueAtTime(
+      this.muteGainNode.gain.value,
+      this.audioContext.currentTime,
+    );
+    this.muteGainNode.gain.linearRampToValueAtTime(
+      1,
+      this.audioContext.currentTime + 0.01,
+    );
   };
 
   /**